Split Buildvana.Core.Json into IJsonHelper contract and impl
Move the JSON helpers behind a new IJsonHelper interface (in
Buildvana.Core.Abstractions) so consumers can be unit-tested without
reaching the file system or System.Text.Json. The concrete JsonHelper
holds an IBuildHost via constructor injection; the public delegate
JsonStringValueRewriter moves to abstractions alongside the contract.

Buildvana.Tool consumers (VersionService, VersionFile,
SelfReferenceUpdater, ReleaseTask) inject and forward IJsonHelper;
BuildContext registers it as a singleton.
